refactor(app): move drag and noDrag to common
