+++ packages/aoboshi-app/src/renderer/app/Toolbar/Toolbar.css.ts
@@ -1,10 +1,10 @@
 import { style } from "@vanilla-extract/css";
-import { windowControlsHeight } from "~/renderer/styles.css";
+import * as theme from "~theme/theme.css";
 
 export const toolbar = style({
   display: "flex",
   alignItems: "center",
-  height: windowControlsHeight,
+  height: theme.vars.windowControls.height,
   transitionProperty: "padding-left",
   transitionDuration: "600ms",
   transitionTimingFunction: "cubic-bezier(.2,0,0,1)",
