+++ packages/aoboshi-app/src/renderer/app/Toolbar/Toolbar.tsx
@@ -1,9 +1,9 @@
 import { ComponentPropsWithoutRef, FunctionComponent } from "react";
 import * as styles from "./Toolbar.css";
-import { windowControlsWidth } from "~/renderer/styles.css";
 import { Container } from "~common/Container";
 import { selectSidebarOpen } from "~app/settingsSlice";
 import { useSelector } from "~app/useSelector";
+import * as theme from "~theme/theme.css";
 
 type ToolbarProps = ComponentPropsWithoutRef<"div">;
 
@@ -19,7 +19,9 @@ export const Toolbar: FunctionComponent<ToolbarProps> = ({
       className={styles.toolbar}
       style={{
         ...style,
-        paddingLeft: sidebarOpen ? 0 : `calc(${windowControlsWidth} + 28px)`,
+        paddingLeft: sidebarOpen
+          ? 0
+          : `calc(${theme.vars.windowControls.width} + 28px)`,
       }}
       {...props}
     >
