+++ packages/aoboshi-app/src/renderer/common/window.css.ts
@@ -1,0 +1,13 @@
+import { style } from "@vanilla-extract/css";
+import { commonLayer } from "~/renderer/layers.css";
+
+export const drag = style(
+  commonLayer({
+    WebkitAppRegion: "drag",
+  }),
+);
+export const noDrag = style(
+  commonLayer({
+    WebkitAppRegion: "no-drag",
+  }),
+);
